+++ src/test/psdReader.spec.ts
@@ -47,6 +47,7 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readFilesPath).filter(f => !/pattern/.test(f)).forEach(f => {
+	// fs.readdirSync(readFilesPath).filter(f => /artboards/.test(f)).forEach(f => {
 		it(`reads PSD file (${f})`, () => {
 			const basePath = path.join(readFilesPath, f);
 			const psd = readPsdFromFile(path.join(basePath, 'src.psd'), { ...opts });
@@ -115,13 +116,14 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readWriteFilesPath).forEach(f => {
+	// fs.readdirSync(readWriteFilesPath).filter(f => /shapes/.test(f)).forEach(f => {
 		it(`reads-writes PSD file (${f})`, () => {
 			const psd = readPsdFromFile(path.join(readWriteFilesPath, f, 'src.psd'), { ...opts, useImageData: true, useRawThumbnail: true });
 			const actual = writePsdBuffer(psd, { logMissingFeatures: true });
 			const expected = fs.readFileSync(path.join(readWriteFilesPath, f, 'expected.psd'));
 			fs.writeFileSync(path.join(resultsFilesPath, `read-write-${f}.psd`), actual);
 
-			// fs.writeFileSync(path.join(resultsFilesPath, `read-write-${f}.bin`), actual);
+			fs.writeFileSync(path.join(resultsFilesPath, `read-write-${f}.bin`), actual);
 			// console.log('------------');
 			// readPsdFromFile(path.join(resultsFilesPath, `read-write-${f}.psd`), { ...opts, useImageData: true, useRawThumbnail: true });